django/requirements.txt
Django==1.6.5
DjangoRestless==0.0.9
Pillow==2.5.1
South==1.0
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
celery==3.1.15
django-admin-bootstrapped==1.6.6
django-celery==3.1.16
django-tastypie==0.12.0
gunicorn==19.1.0
kombu==3.0.23
numpy==1.8.1
python-dateutil==2.2
python-mimeparse==0.1.4
pytz==2014.7
six==1.8.0
sorl-thumbnail==11.12
wsgiref==0.1.2
